ZipRecruiter - Update 0.14.0 InDev
Added ZipRecruiter functionality to the addon. It now will rescan when a DOM change is detected, and will act with a delay as to not make the user's page flicker with the highlighted spans.

Revised approach to Semantic Versioning to match behavior expected of a program in development. Revised from version 3.0.1 down to 0.14.0 to indicate the development state of the addon. Significant strides have been made since the proof of concept, and the addon is approaching a more launchable state by the day.