Abstracted stateScores, changed text processing
stateScores are now a window in their own human readable file and are no longer contained within content.js as a minified object

Text is now processed in a way that grabs preceding and trailing text from the target state to be evaluated. This does not fix the bug with Indeed having the car removed, that is a different problem to be fixed in another build (it is an html element and not text).

Search behavior across coloring functions is now consistent.