v0.19.1 Regular Expressions
Replaced the ever expanding matching logic to simply match the key values from stateScores.js using regular expressions. More and more dev time was being dedicated to accounting for all the different ways states could be listed, and it was too large a drain. This regex, while a more blunt instrument than the granularity of the many if statements, is also a whole lot simpler to maintain.