Fixed linkedin.com/jobs/ page
The basic jobs page (not search) was not being searched correctly. Added a second querySelectorAll that calls all artdeco-entity-lockup__subtitle divs with a child div that has the property div[dir="ltr"].

DOM monitoring on this page is still an issue. Reprocessing the spans works, so it is a DOM monitor issue.