Fixed LinkedIn Observer & /jobs/ page
The observer for LinkedIn was not monitoring for DOM changes on the linkedin.com/jobs/ page.  Added parameters for the double div setup for DOM observer (/jobs/* is div then span by comparison).

Updated the logic for linkedin's coloring function to check if it is on a /jobs/ page or a /jobs/* (search) page to more efficiently track content.