File reorganization
Changed Manifest.json to have a more readable format.

Changed Manifest.json to reflect changes in the file structure of the program and adhere to manifest standards.

Changed the file structure of the extension by moving the icons into a dedicated icon folder.

Changed the file structure of the extension by moving the popup and related files into a dedicated browserAction folder.