Dynamic popup tooltips, fixed popup icon
Fixed an issue where the filepath was not correctly specified for the icon in popup.html.

Changed popup.html to have id tags associated with the div to allow for dynamic tooltip changes and also allow for different languages to be displayed in the text fields.

Added a function to popup.js that runs on load and on change of mode that dynamically changes the popups to be associated with the respective scores, as the thresholds for gender identity and overall policy are different.